begin enforce support
